requirements.txt
Flask==2.0.2
blinker==1.4
watchdog==2.1.2
lxml==4.6.3
greenlet==1.1.0
Cython==0.29.23
brotlipy==0.7.0
img2pdf==0.4.1
importlib_metadata==3.10.0
requests==2.27.1
numpy==1.20.2
ocrmypdf==12.1.0
gevent==21.1.2
tornado==6.1
keyring==23.0.1
zipp==3.4.1
typing_extensions==3.7.4.3
docutils==0.17.1
cryptography==3.4.7
dataclasses==0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
brotli==1.0.9
contextvars==2.4
lockfile==0.12.2
monotonic==1.6
ordereddict==1.1
Pillow==8.3.2
protobuf==3.18.0
pyOpenSSL==20.0.1
python-dotenv==0.19.0
wincertstore
urllib3==1.25.9
Werkzeug==2.0.2
gunicorn==20.1.0

